+++ punto_back/tests/games.test.js
@@ -3,13 +3,25 @@ const request = require("supertest");
 const app = require('../server')
 const Game = require('../models/Game')
 const Board = require('../models/Board')
-const Card = require('../models/Card')
 const User = require('../models/User')
+const jwt = require("jsonwebtoken");
+
+const jwtToken = jwt.sign(
+    {
+        "UserInfo": {
+            "email": 'farinel.sacha@gmail.com'
+        }
+    },
+    process.env.ACCESS_TOKEN_SECRET,
+    {expiresIn: '15m'}
+)
 
 // Connecting to the database before each test.
 beforeAll(async () => {
     await mongoose.connect(process.env.DATABASE_URI_TEST);
-    await mongoose.connection.dropCollection("games")
+    if (mongoose.connection.collections.games) {
+        await mongoose.connection.dropCollection("games");
+    }
 });
 
 // Closing database connection after each test.
@@ -19,7 +31,9 @@ afterAll(async () => {
 
 describe("GET /games on empty {collection}", () => {
     it("should return 'No games found'", async () => {
-        const res = await request(app).get("/games");
+        const res = await request(app)
+            .get("/games")
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(400);
         expect(res.body.message).toBe('No games found');
     });
@@ -36,23 +50,28 @@ describe("POST /games", () => {
             birthday: "27/10/1994"
         });
         const userID = (user._id).toString()
-        const res = await request(app).post("/games").send({
-            board: boardID,
-            hands:
-                [
-                    {
-                        idPlayer: userID,
-                        cards: []
-                    }
-                ]
-        });
+        const res = await request(app)
+            .post("/games")
+            .send({
+                board: boardID,
+                hands:
+                    [
+                        {
+                            idPlayer: userID,
+                            cards: []
+                        }
+                    ]
+            })
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(201);
     });
 });
 
 describe("GET /games", () => {
     it("should return all boards", async () => {
-        const res = await request(app).get("/games");
+        const res = await request(app)
+            .get("/games")
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(200);
         expect(res.body.length).toBeGreaterThan(0);
     });
@@ -80,7 +99,9 @@ describe("GET /games/:id", () => {
                 ]
         })
         const gameID = ((game._id).toString())
-        const res = await request(app).get("/games?id=" + gameID);
+        const res = await request(app)
+            .get("/games?id=" + gameID)
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(200);
     });
 });
@@ -107,9 +128,10 @@ describe("DELETE /games", () => {
                 ]
         })
         const gameID = ((game._id).toString())
-        const res = await request(app).delete("/games").send({
-            id: gameID
-        });
+        const res = await request(app)
+            .delete("/games")
+            .send({id: gameID})
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(200);
     });
 });
