+++ punto_back/tests/users.test.js
@@ -3,11 +3,24 @@ const request = require("supertest");
 const app = require('../server')
 const User = require('../models/User')
 const connectDB = require("../config/dbConn");
+const jwt = require("jsonwebtoken");
+
+const jwtToken = jwt.sign(
+    {
+        "UserInfo": {
+            "email": 'farinel.sacha@gmail.com'
+        }
+    },
+    process.env.ACCESS_TOKEN_SECRET,
+    {expiresIn: '15m'}
+)
 
 // Connecting to the database before each test.
 beforeAll(async () => {
     connectDB(process.env.DATABASE_URI_TEST);
-    await mongoose.connection.dropCollection("users")
+    if (mongoose.connection.collections.users) {
+        await mongoose.connection.dropCollection("users");
+    }
 });
 
 // Closing database connection after each test.
@@ -17,7 +30,9 @@ afterAll(async () => {
 
 describe("GET /users on empty {collection}", () => {
     it("should return 'No users found'", async () => {
-        const res = await request(app).get("/users");
+        const res = await request(app)
+            .get("/users")
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(400);
         expect(res.body.message).toBe('No users found');
     });
@@ -25,11 +40,13 @@ describe("GET /users on empty {collection}", () => {
 
 describe("POST /users without all fields", () => {
     it("should return 400 'All fields are required'", async () => {
-        const res = await request(app).post("/users").send({
-            email: "farinel.sacha@gmail.com",
-            username: "SachaFARINEL",
-            password: "test",
-        });
+        const res = await request(app)
+            .post("/users")
+            .send({
+                email: "farinel.sacha@gmail.com",
+                username: "SachaFARINEL",
+                password: "test",
+            });
         expect(res.statusCode).toBe(400);
         expect(res.body.message).toBe("All fields are required");
 
@@ -38,12 +55,14 @@ describe("POST /users without all fields", () => {
 
 describe("POST /users", () => {
     it("should create a user", async () => {
-        const res = await request(app).post("/users").send({
-            email: "farinel.sacha@gmail.com",
-            username: "SachaFARINEL",
-            password: "test",
-            birthday: "27/10/1994"
-        });
+        const res = await request(app)
+            .post("/users")
+            .send({
+                email: "farinel.sacha@gmail.com",
+                username: "SachaFARINEL",
+                password: "test",
+                birthday: "27/10/1994"
+            });
         expect(res.statusCode).toBe(201);
         expect(res.body.message).toBe("New user farinel.sacha@gmail.com created");
 
@@ -52,12 +71,14 @@ describe("POST /users", () => {
 
 describe("POST duplicate /users", () => {
     it("should return 'Duplicate username' error", async () => {
-        const res = await request(app).post("/users").send({
-            email: "farinel.sacha@gmail.com",
-            username: "SachaFARINEL",
-            password: "test",
-            birthday: "27/10/1994"
-        });
+        const res = await request(app)
+            .post("/users")
+            .send({
+                email: "farinel.sacha@gmail.com",
+                username: "SachaFARINEL",
+                password: "test",
+                birthday: "27/10/1994"
+            });
         expect(res.statusCode).toBe(409);
         expect(res.body.message).toBe("Duplicate username");
 
@@ -66,7 +87,9 @@ describe("POST duplicate /users", () => {
 
 describe("GET /users", () => {
     it("should return all users", async () => {
-        const res = await request(app).get("/users");
+        const res = await request(app)
+            .get("/users")
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(200);
         expect(res.body.length).toBeGreaterThan(0);
     });
@@ -75,9 +98,10 @@ describe("GET /users", () => {
 describe("DELETE /user", () => {
     it("should delete a user", async () => {
         const user = await User.findOne({email: 'farinel.sacha@gmail.com'}).lean().exec()
-        const res = await request(app).delete("/users").send({
-            id: user._id.toString()
-        });
+        const res = await request(app)
+            .delete("/users")
+            .send({id: user._id.toString()})
+            .set('Authorization', `Bearer ${jwtToken}`);
         expect(res.statusCode).toBe(200);
     });
 });
